+++ shared/src/led/led.hpp
@@ -12,16 +12,16 @@ namespace SyncBlink
     class LED
     {
     public:
-        void setup(uint32_t ledCount);
+        void setup(uint16_t ledCount);
         void loop();
 
         void setAllLeds(const Color color);
 
-        void setLed(const uint32_t index, const Color color);
-        uint32_t getLed(const uint32_t index);
-        uint32_t getLedCount() const;
+        void setLed(const uint16_t index, const Color color);
+        uint16_t getLed(const uint16_t index);
+        uint16_t getLedCount() const;
 
-        void setGroup(const uint32_t index, std::vector<uint32_t> group);
+        void setGroup(const uint16_t index, std::vector<uint16_t> group);
         void clearGroups();
 
         void showNow(const Color color);
@@ -32,7 +32,7 @@ namespace SyncBlink
         bool _groupsSet = false;
 
         std::vector<CRGB> _leds;
-        std::vector<std::vector<uint32_t>> _groups;
+        std::vector<std::vector<uint16_t>> _groups;
     };
 }
